+++ src/services/appearance.js
@@ -1,8 +1,13 @@
 /* global CONFIG */
 import storage from 'local-storage-fallback';
+import { isMobile } from './is-mobile';
 
-export const { colorSchemeStorageKey, nsfwVisibilityStorageKey, uiScaleStorageKey } =
-  CONFIG.appearance;
+export const {
+  colorSchemeStorageKey,
+  nsfwVisibilityStorageKey,
+  uiScaleStorageKey,
+  submitModeStorageKey,
+} = CONFIG.appearance;
 
 export const SCHEME_LIGHT = 'light';
 export const SCHEME_DARK = 'dark';
@@ -63,4 +68,22 @@ export function saveUIScale(scale) {
     // Default value
     storage.removeItem(uiScaleStorageKey);
   }
+}
+
+export function loadSubmitMode() {
+  const mode = storage.getItem(submitModeStorageKey);
+  return ['enter', 'ctrl+enter', 'auto'].includes(mode) ? mode : 'auto';
+}
+
+export function saveSubmitMode(mode) {
+  if (['enter', 'ctrl+enter'].includes(mode)) {
+    storage.setItem(submitModeStorageKey, mode);
+  } else {
+    // Default value
+    storage.removeItem(submitModeStorageKey);
+  }
+}
+
+export function submittingByEnter(submitMode) {
+  return submitMode === 'enter' || (submitMode === 'auto' && !isMobile);
 }