+++ test/jest/post.test.js
@@ -47,12 +47,16 @@ const defaultState = {
   sendTo: { feeds: [{ id: 'feed-id', user: AUTHOR }] },
   serverInfoStatus: successAsyncState,
   serverInfo: { features: { perGroupsPostDelete: true } },
+  submitMode: 'enter',
 };
 
 const renderPost = (props = {}, options = {}) => {
   const { Provider } = reactRedux;
   const dummyReducer = (state) => state;
-  const store = createStore(dummyReducer, defaultState);
+  const store = createStore(dummyReducer, {
+    ...defaultState,
+    submitMode: options.submitMode || 'enter',
+  });
 
   const defaultProps = {
     id: 'post-id',
@@ -265,9 +269,7 @@ describe('Post', () => {
   });
 
   it('Renders a like button which likes the post', () => {
-    const someOtherUser = {
-      id: 'other-id',
-    };
+    const someOtherUser = { id: 'other-id' };
     const likePost = jest.fn();
     renderPost({ likePost, isEditable: false, user: someOtherUser });
     expect(screen.getByText('Like', { role: 'button' })).toBeInTheDocument();
@@ -276,9 +278,7 @@ describe('Post', () => {
   });
 
   it('Renders an un-like button which un-likes the post if this post is already liked', () => {
-    const someOtherUser = {
-      id: 'other-id',
-    };
+    const someOtherUser = { id: 'other-id' };
     const unlikePost = jest.fn();
     renderPost({
       unlikePost,
@@ -376,9 +376,7 @@ describe('Post', () => {
   });
 
   it('Renders a hide button which hides the post', () => {
-    const someOtherUser = {
-      id: 'other-id',
-    };
+    const someOtherUser = { id: 'other-id' };
     const hidePost = jest.fn();
     renderPost({
       user: someOtherUser,
@@ -392,9 +390,7 @@ describe('Post', () => {
   });
 
   it('Renders a un-hide button which unhides the post', () => {
-    const someOtherUser = {
-      id: 'other-id',
-    };
+    const someOtherUser = { id: 'other-id' };
     const unhidePost = jest.fn();
     renderPost({
       user: someOtherUser,
@@ -422,11 +418,72 @@ describe('Post', () => {
     expect(cancelEditingPost).toHaveBeenCalledWith('post-id');
   });
 
-  it('Lets me edit text of my post by typing', () => {
+  it('Lets me edit text of my post by typing with Shift+Enter when "submitMode" is "enter"', () => {
     const saveEditingPost = jest.fn();
-    renderPost({ isEditing: true, isEditable: true, saveEditingPost });
+    renderPost({
+      isEditing: true,
+      isEditable: true,
+      saveEditingPost,
+    });
 
     userEvent.type(screen.getByRole('textbox'), 'Hello,{shift}{enter}{/shift}World!{enter}');
+    expect(screen.getByRole('textbox')).toHaveValue('Hello,\nWorld!');
+    expect(saveEditingPost).toHaveBeenCalledWith('post-id', {
+      attachments: [],
+      body: 'Hello,\nWorld!',
+      feeds: ['author'],
+    });
+  });
+
+  it('Lets me edit text of my post by typing with Alt+Enter when "submitMode" is "enter"', () => {
+    const saveEditingPost = jest.fn();
+    renderPost({
+      isEditing: true,
+      isEditable: true,
+      saveEditingPost,
+    });
+
+    userEvent.type(screen.getByRole('textbox'), 'Hello,{alt}{enter}{/alt}World!{enter}');
+    expect(screen.getByRole('textbox')).toHaveValue('Hello,\nWorld!');
+    expect(saveEditingPost).toHaveBeenCalledWith('post-id', {
+      attachments: [],
+      body: 'Hello,\nWorld!',
+      feeds: ['author'],
+    });
+  });
+
+  it('Lets me submit text of my post by Ctrl+Enter typing when "submitMode" is "ctrl+enter"', () => {
+    const saveEditingPost = jest.fn();
+    renderPost(
+      {
+        isEditing: true,
+        isEditable: true,
+        saveEditingPost,
+      },
+      { submitMode: 'ctrl+enter' },
+    );
+
+    userEvent.type(screen.getByRole('textbox'), 'Hello,{enter}World!{ctrl}{enter}{/ctrl}');
+    expect(screen.getByRole('textbox')).toHaveValue('Hello,\nWorld!');
+    expect(saveEditingPost).toHaveBeenCalledWith('post-id', {
+      attachments: [],
+      body: 'Hello,\nWorld!',
+      feeds: ['author'],
+    });
+  });
+
+  it('Lets me submit text of my post by Meta+Enter typing when "submitMode" is "ctrl+enter"', () => {
+    const saveEditingPost = jest.fn();
+    renderPost(
+      {
+        isEditing: true,
+        isEditable: true,
+        saveEditingPost,
+      },
+      { submitMode: 'ctrl+enter' },
+    );
+
+    userEvent.type(screen.getByRole('textbox'), 'Hello,{enter}World!{meta}{enter}{/meta}');
     expect(screen.getByRole('textbox')).toHaveValue('Hello,\nWorld!');
     expect(saveEditingPost).toHaveBeenCalledWith('post-id', {
       attachments: [],